feat: 7단계
