+++ src/main/java/roomescape/auth/JwtTokenProvider.java
@@ -1,0 +1,43 @@
+package roomescape.auth;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.security.Keys;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import roomescape.member.MemberResponse;
+
+import javax.crypto.SecretKey;
+import java.nio.charset.StandardCharsets;
+
+@Configuration
+public class JwtTokenProvider {
+    @Value("${roomescape.auth.jwt.secret}")
+    private String secretKey;
+
+    @Bean
+    public SecretKey secretKey() {
+        return Keys.hmacShaKeyFor(secretKey.getBytes(StandardCharsets.UTF_8));
+    }
+
+    public String createToken(MemberResponse member) {
+        String accessToken = Jwts.builder()
+                .setSubject(member.getId().toString())
+                .claim("name", member.getName())
+                .claim("email", member.getEmail())
+                .signWith(secretKey())
+                .compact();
+
+        return accessToken;
+    }
+
+    public Long extractMemberIdFromToken(String token) {
+        Claims claims = Jwts.parserBuilder()
+                .setSigningKey(secretKey())
+                .build()
+                .parseClaimsJws(token)
+                .getBody();
+        return Long.valueOf(claims.getSubject());
+    }
+}
