+++ src/main/java/roomescape/member/MemberController.java
@@ -3,26 +3,18 @@ package roomescape.member;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-import roomescape.instructure.JwtTokenProvider;
+import org.springframework.web.bind.annotation.*;
 
 import java.net.URI;
 
 @RestController
 public class MemberController {
-    private JwtTokenProvider jwtUtil;
     private MemberService memberService;
 
-    public MemberController(MemberService memberService, JwtTokenProvider jwtUtil) {
+    public MemberController(MemberService memberService) {
         this.memberService = memberService;
-        this.jwtUtil = jwtUtil;
     }
 
     @PostMapping("/members")
@@ -34,7 +26,7 @@ public class MemberController {
     @PostMapping("/login")
     public ResponseEntity<String> login(@RequestBody MemberRequest memberRequest, HttpServletResponse response){
         MemberResponse member = memberService.findMember(memberRequest.getEmail(), memberRequest.getPassword());
-        if (member == null) { //멤버 X
+        if (member == null) {
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         }
         String token = memberService.createToken(member);
@@ -43,14 +35,12 @@ public class MemberController {
         return ResponseEntity.status(HttpStatus.OK).build();
     }
 
-
     @GetMapping("/login/check")
     public ResponseEntity<MemberResponse> checkLogin(HttpServletRequest request) {
         Cookie[] cookies = request.getCookies();
         String token = memberService.extractTokenFromCookie(cookies);
 
-        Long memberId = memberService.extractMemberIdFromToken(token);
-        MemberResponse member = memberService.findMemberById(memberId);
+        MemberResponse member = memberService.findMemberById(memberService.extractMemberFromToken(token).getId());
 
         MemberResponse memberResponse = new MemberResponse(member.getId(), member.getName(), member.getEmail());
         return ResponseEntity.ok().body(memberResponse);
