+++ src/main/java/roomescape/member/MemberService.java
@@ -1,14 +1,10 @@
 package roomescape.member;
 
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
-import roomescape.instructure.JwtTokenProvider;
+import roomescape.auth.JwtTokenProvider;
 
 import java.util.Optional;
 
@@ -23,9 +19,6 @@ public class MemberService {
         this.memberRepository = memberRepository;
         this.jwtTokenProvider = jwtTokenProvider;
     }
-
-    @Value("${roomescape.auth.jwt.secret}")
-    String secretKey;
 
     public MemberResponse createMember(MemberRequest memberRequest) {
         Member member = memberRepository.save(new Member(memberRequest.getName(), memberRequest.getEmail(), memberRequest.getPassword(), "USER"));
@@ -56,21 +49,12 @@ public class MemberService {
     }
 
     public Member extractMemberFromToken(String token) {
-        Long memberId = extractMemberIdFromToken(token);
+        Long memberId = jwtTokenProvider.extractMemberIdFromToken(token);
         if (memberId != null) {
             Optional<Member> optionalMember = memberRepository.findById(memberId);
             return optionalMember.orElse(null);
         }
         return null;
-    }
-
-    public Long extractMemberIdFromToken(String token) {
-        Claims claims = Jwts.parserBuilder()
-                .setSigningKey(Keys.hmacShaKeyFor(secretKey.getBytes()))
-                .build()
-                .parseClaimsJws(token)
-                .getBody();
-        return Long.valueOf(claims.getSubject());
     }
 
     public String extractTokenFromCookie(Cookie[] cookies) {
