+++ src/test/java/roomescape/MissionStepTest.java
@@ -6,8 +6,10 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.stereotype.Component;
 import org.springframework.test.annotation.DirtiesContext;
 
+import roomescape.auth.JwtTokenProvider;
 import roomescape.reservation.MyReservationResponse;
 import roomescape.reservation.ReservationResponse;
 import roomescape.waiting.WaitingResponse;
@@ -173,4 +175,9 @@ public class MissionStepTest {
         assertThat(status).isEqualTo("1번째 예약대기");
     }
 
+    @Test
+    void 칠단계() {
+        Component componentAnnotation = JwtTokenProvider.class.getAnnotation(Component.class);
+        assertThat(componentAnnotation).isNull();
+    }
 }
